Cleaned up code, added full comments. FINAL
This commit took a while, and I'm a little disappointed to be wrapping
up the project like this. I was not able to figure out how to refactor
my code to allow for local storage functionality. I am really proud of
how the final project turned out outside of that though. I could've used
a mentor for localstorage, I read everything I could on it and watched
videos but couldn't get it to work. I found myself not looking forward
to working on this class, and it became more of a chore. So in the
interest of this being a hobby at this point, I am going to move on and
continue my learning journey.